'super'comment out the old colours.
Ruby or jekyll or whatever has a ridiculous yaml parser that still slurps up
comments to send to the liquid templating engine and we end up with corrupt css
files. This commit fixes that.

Sadly index.html's ids don't manage to correctly trigger the #20150102topic etc
selectors for some reason.